p2p_project/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
cryptography==41.0.3
Django==4.2.2
django-cachalot==2.6.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-paypal==2.0
django-recaptcha==3.0.0
django-tables2==2.6.0
Django-Verify-Email==2.0.3
django-widget-tweaks==1.4.12
gino==1.0.1
gunicorn==21.2.0
hiredis==2.2.3
idna==3.4
packaging==23.1
paramiko==3.3.1
psycopg2==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.3.24
sqlparse==0.4.4
sshtunnel==0.4.0
urllib3==2.0.6
whitenoise==6.5.0
